resolves #4
Both server and client accept a :tls => false, true, {} option. If false
then TLS is not used. If true then TLS is used with the default
EventMachine start_tls options. If a Hash is passed then it is used to
override the default options to start_tls.

The server banner includes the scheme - pryem, or pryems - that the
server is expecting. If the client was started in a mode that doesn't
match the server's banner scheme then the connection is terminated and
an error is displayed.

NOT TESTED: what happens if TLS negotiation fails? Is the TLS traffic
actually secure when the default options are used?