closes #5 - adds user authentication support
When creating the service pass anything that responds to #call,
or #[](name), e.g., a Hash, in the :auth option. If it's a Hash
the keys should be usernames and the values shoud be passwords.
If you pass a Proc or anything that responds to #call it should
take two arguments (name, password) and return either true or
false. Nil is probably not handled well.

When a client connects it will authenticate itself if it was created
with :user and :pass options. If the server is not configured with the
:auth option then all auth requests will return success (true).

When a server receives data from a client - after initial banner and
TLS negotiation - it will demand the client authenticate if the server
was started with the :auth option. The server will not process any other
type of data from the client until authentication has completed.

If the client provides invalid authentication information 5 times the
server will terminate the connection after sending an error message.

In the future the :pass option should be allowed to be a Proc that can
be called each time the server rejects the authentication attempt.