refs #9 #5 tab completion doesn't raise yield from root Fiber errors
Authentication is now strictly initiated by the server. If the client
attempts to authenticate before the server demands it then a race
condition could occur. It does ensure that the client will only ask the
user to enter credentials if they are requried by the server.

A suite of regression tests really is needed...