refs #25 Adds simple chat with '!' and '!!'
Any line read from the client that starts with '!' or '!!' will be sent
to other clients as a message for the user. ! messages will be sent to
all connections for the same object. !! messages will be sent to all
users of the current Process.

The client will not display the message until the users presses enter.
If there is a message to display it will be written to STDOUT and
highlighted. It will also be preceeded with the marker for the type of
message it was. There is no support for one-to-one messages. There is no
support for identifying the sender of the message.

The client side still blocks on readline. Ultimately I'd like the client
to be non blocking and display message as they are received in a
dedicated area of the terminal.