json proto is abstracted and more robust
client and server don't directly send and receive json; the proto
defines simple prototypes for the types of messages that can be sent and
received. the client and server then create specific implementations for
each type of message

the json proto doesn't use a simple string delemiter between json
payloads. instead each payload is prefixed with a crc, length, etc.,.
i'd rather the length and crc fields be packed, but JS implementations
of pack and unpack aren't straight forward. there are currently no JS
implementations, but i want to be able to provide one in the future.