adds a Broker to keep track of all listening servers on this host
by default the client will connect to the broker and present the user
with a list of avaialable pry remote servers. the user can select one by
id, name, or url. once the user selects an available server the
connection to the broker will be closed and a new connection to the
selected sever will be established. the client can still connect directly
to servers without first going through the broker.  if the broker stops
one of it's clients will eventually restart it. servers must refresh
their registration every 20 seconds or be unregistered

currently the client is only known to be compatible with brokers running
on the local machine.

provides: PryRemoteEm.{servers, stop_server}; register with restarted Brokers quickly

changes: when specifying a specific port to listen on the option
:port_fail can also be specified. if the specified port is already in
use and :port_fail is :auto the next available port will be used instead.