WIP: client will work with broker
by default the client will connect to the broker and present the user
with a list of avaialable pry remote servers. the user can select one by
id, name, or url. once the user selects an available server the
connection to the broker will be closed and a new connection to the
selected sever will be established.

currently the client is only known to be compatible with brokers running
on the local machine.

the client can still connect directly to servers without first going
through the broker.