sort and filter; accept server name to proxy or connect to on command line
the server list can be sorted by server name, server host, server port
ssl support. it can also be filterd based on the same four attributes.

specifying the -c, or -p option will cause the client to search the
server list for a matching server and immediately connect, or proxy to
it.