adds support for edit[-*] commands; Adds Pry#_pryem_ attribute
file contents will be sent to the client that requested the edit. the
client will load the contents in an editor. once the editor quits the
new contents will be sent back to the server. if the file was changed
before the client returned the new contents the client will be warned
and asked to confirm that the file should be overwritten.

Pry instances that are created as part of a PryRemoteEm server will
respond to Pry#_pryem_ with the client connection. Pry helpers, etc.,.
can use the presnce of that attribute to determine if they are being
executed by a remote connection. They can also use that attribute to
communicate with the client.